+++ spec/support/models.rb
@@ -156,5 +156,35 @@ module ModelsHelper
     end
   end
 
+  def load_belongs_to_polymorphic_models
+    define_class :Logo do
+      include NoBrainer::Document
+
+      belongs_to :imageable, polymorphic: true
+    end
+
+    define_class :Picture do
+      include NoBrainer::Document
+
+      belongs_to :imageable, polymorphic: true
+    end
+
+    define_class :Event do
+      include NoBrainer::Document
+
+      belongs_to :restaurant
+      has_many :photos, as: :imageable, class_name: 'Picture'
+    end
+
+    define_class :Restaurant do
+      include NoBrainer::Document
+
+      has_one :logo, as: :imageable
+      has_many :pictures, as: :imageable
+      has_many :events
+      has_many :photos, through: :events
+    end
+  end
+
   RSpec.configure { |config| config.include self }
 end